Rename file from knex.js to adapter.js
