+++ src/database/adapter.js
